create grid layout component
