+++ components/GridLayout/GridLayout.module.scss
@@ -1,0 +1,24 @@
+.content {
+    padding-top: 20vh;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    gap: 1rem;
+    z-index: 1;
+}
+
+.content h1 {
+    font-size: 65px;
+    z-index: 1;
+}
+
+.body {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: space-between;
+    align-items: center;
+    z-index: 1;
+    padding: 4rem;
+    width: 100%;
+}
